Edge Detection
In this project, the initial step is to read an image and convert it to grayscale. The technique employed in Python is to evaluate adjacent pixels in the image and pinpoint locations where there are substantial variations in pixel values. These pixels are considered edge pixels if they surpass a specified threshold, and they are then plotted.